day4/workspace/requirements.txt
dnspython==2.8.0
mysql-connector-python==9.4.0
pymongo==4.15.1
